Change the GHA workflows so releases can be automated
The release process currently requires a couple of manual steps that is
not documented completely. A new workflow has been added to prepare the
release and the existing build workflow has been refactored so that it
can create the GitHub release as a part of the build. Documentation has
been added to reflect these new changes which should hopefully make it a
lot easier for someone new to the project to create a release.

Signed-off-by: Remy Suen <remy.suen@docker.com>