Handle inlay hint requests asynchronously
Inlay hints on Dockerfiles need to fetch data so it can take some time
before the result can be returned. This means that everything else gets
blocked because all JSON-RPC messages are being handled synchronously at
the moment. To alleviate this problem, we will check for inlay hint
requests and then handle them in a goroutine asynchronously so that
other messages can continue to be processed.

Other messages may be changed to be handled asynchronously in the future
but only inlay hint requests will be done this way for the time being.
We want to be careful about introducing drastic changes to how messages
are being handled as it is critical for the operation of the language
server.

Signed-off-by: Remy Suen <remy.suen@docker.com>